@zpoint 2024-11-19
Hi @romilbhardwaj, this [PR](https://github.com/zpoint/skypilot/pull/1) adds more test cases and global functions based on current PR.

To avoid disrupting what's already reviewed, I opened another [PR](https://github.com/zpoint/skypilot/pull/1) based on this one. Could you please help review it? Thanks!